+++ res/views/components/button.blade.php
@@ -4,7 +4,7 @@ $href = isset($route) ? route($route, $params) : ($href ? url($href) : null);
 @endphp
 
 @isset($href)
-    @ui('a', $attributes->merge(compact('href')), $slot)
+    @tag('a', $attributes->merge(compact('href')), $slot)
 @else
-    @ui('button', $attributes->merge(compact('type')), $slot)
+    @tag('button', $attributes->merge(compact('type')), $slot)
 @endisset