+++ res/views/components/link.blade.php
@@ -2,8 +2,8 @@
 $href = isset($route) ? route($route, $params) : url($href);
 @endphp
 
-@ui('a', $attributes->merge(compact('href')))
+@tag('a', $attributes->merge(compact('href')))
     <s-text :value="$text">
         {{ $slot }}
     </s-text>
-@endui
+@endtag